extract nic-emu example from simbricks/**/nicbm.h
